backend/app/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cerebras_cloud_sdk==1.50.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
fastapi==0.118.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
numpy==2.3.3
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
